neurips21/requirements.txt
numpy==1.16.5
tensorflow_gpu==1.15.0
keras==2.2.4
matplotlib==3.0.3
pandas==0.24.2
scipy==1.2.1
seaborn==0.9.0
tabulate==0.8.3
